requirements.txt
# --- Core Dependencies ---
# Эти библиотеки необходимы для запуска Vintix Corobot.
# Версии зафиксированы для обеспечения воспроизводимости.

# Компьютерное зрение и обработка данных
opencv-python==4.9.0.80
numpy==1.26.4
pyyaml==6.0.1
pillow==10.3.0
h5py==3.11.0
scipy==1.13.1
matplotlib==3.8.4

# Робототехника и кинематика
ikpy==3.4.2 # Обновлено до последней доступной версии

# Нейросеть Vintix
torch==2.3.0
huggingface-hub==0.23.0

# Распознавание речи
vosk==0.3.45
pyaudio==0.2.14

# Сетевое взаимодействие
requests==2.31.0

# Опциональные зависимости
# psutil>=5.8.0  # для мониторинга системных ресурсов
# tqdm>=4.56.0  # для отображения прогресса операций
